requirements.txt
tkinter
imdbpy
duckduckgo-search
sqlite3
aiohttp